pyproject.toml
[project]
name = "mindtune"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "joblib>=1.5.3",
    "lightgbm>=4.6.0",
    "numpy>=2.4.4",
    "pandas>=3.0.1",
    "pypdf>=6.9.0",
    "pyserial>=3.5",
    "scikit-learn>=1.6.1,<1.7",
]
